feat: implementing try catch and AppError

+++ src/routes/tasks.routes.js
@@ -1,6 +1,7 @@
 import { buildRoutePath } from "../utils/build-route-path.js";
 import { Database } from "../database/index.js";
 import { Task } from '../models/Task.js';
+import { AppError } from "../utils/AppError.js";
 
 const database = new Database();
 
@@ -11,16 +12,31 @@ export const tasksRoutes = [
     handler: (request, response) => {
       const { title, description } = request.body;
 
-      const task = new Task(title, description);
+      try {
+        if (!title) throw new AppError('Title is required', 404);
+        if (!description) throw new AppError('Description is required', 404);
 
-      database.insert('tasks', task);
+        const task = new Task(title, description);
+        database.insert('tasks', task);
 
-      return response.writeHead(202).end();
+      } catch (error) {
+        if (error instanceof AppError) {
+          return response.writeHead(error.statusCode).end(
+            JSON.stringify({ error: error.message })
+          );
+        }
+
+        return response.writeHead(500).end(
+          JSON.stringify({ error: error.message })
+        );
+      }
+
+      return response.writeHead(201).end();
     }
   },
   {
     method: 'GET',
-    path:  buildRoutePath('/tasks'),
+    path: buildRoutePath('/tasks'),
     handler: (request, response) => {
       const { search } = request.query;
 
@@ -40,12 +56,27 @@ export const tasksRoutes = [
       const { title, description } = request.body;
 
       try {
+        const [task] = database.select('tasks', { id });
+
+        if (!task) throw new AppError('Task with id not found', 404);
+
+        if (!title) throw new AppError('Title is required', 404);
+
+        if (!description) throw new AppError('Description is required', 404);
+
         database.update('tasks', id, {
           title,
           description,
         });
       } catch (error) {
-        return response.writeHead(404).end(JSON.stringify({ error: error.message }));
+        if (error instanceof AppError) {
+          return response.writeHead(error.statusCode).end(
+            JSON.stringify({ error: error.message })
+          );
+        }
+        return response.writeHead(500).end(
+          JSON.stringify({ error: error.message })
+        )
       }
 
       return response.writeHead(202).end();
@@ -56,11 +87,22 @@ export const tasksRoutes = [
     path: buildRoutePath('/tasks/:id'),
     handler: (request, response) => {
       const { id } = request.params;
-      
+
       try {
+        const [task] = database.select('tasks', { id });
+
+        if (!task) throw new AppError('Task with id not found', 404);
+
         database.delete('tasks', id);
       } catch (error) {
-        return response.writeHead(404).end(JSON.stringify({ error: error.message }));
+        if (error instanceof AppError) {
+          return response.writeHead(error.statusCode).end(
+            JSON.stringify({ error: error.message })
+          );
+        }
+        return response.writeHead(500).end(
+          JSON.stringify({ error: error.message })
+        )
       }
 
       return response.writeHead(204).end();
@@ -73,9 +115,23 @@ export const tasksRoutes = [
       const { id } = request.params;
 
       try {
-        database.complete('tasks', id);
-      } catch(error) {
-        return response.writeHead(404).end(JSON.stringify({ error: error.message }));
+        const [task] = database.select('tasks', { id });// desetruturando único item do array
+
+        if (!task) throw new AppError('Task with id not found', 404);
+
+        const isTaskCompleted = !!task.completed_at;
+
+        database.update('tasks', id, {
+          completed_at: isTaskCompleted ? null : new Date(),
+        });
+      } catch (error) {
+        if (error instanceof AppError) {
+          return response.writeHead(error.statusCode).end(JSON.stringify({ error: error.message }));
+        }
+
+        return response.writeHead(500).end(
+          JSON.stringify({ error: error.message })
+        )
       }
 
       return response.writeHead(204).end();
